pyproject.toml
[project]
name = "oplatalinkbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.24.0",
    "aiosqlite>=0.20.0",
    "apscheduler>=3.10.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
]
